+++ demo/api/vitest.config.ts
@@ -4,6 +4,7 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
     plugins: [tsconfigPaths()],
     test: {
+        exclude: ["dist/**", "node_modules/**"],
         reporters: ["default", "junit"],
         outputFile: { junit: "./junit.xml" },
     },