+++ packages/admin/admin/src/dateTime/datePickerField/__stories__/DatePickerField.stories.tsx
@@ -171,3 +171,51 @@ export const Disabled: Story = {
         );
     },
 };
+
+/**
+ * Before implementing custom validation, try to disable certain values using the props provided by MUI's DatePicker, such as `shouldDisableDate`, `disableFuture`, `disablePast`.
+ *
+ * Use this when:
+ * - You need to validate the date against a custom rule, which cannot be achieved using the props provided by MUI's DatePicker
+ * - You want to provide a custom error message
+ * - You want to validate the date asynchronously
+ */
+export const CustomValidation: Story = {
+    render: () => {
+        interface FormValues {
+            value: string;
+        }
+
+        const validateIsWeekday = async (value: string | undefined) => {
+            if (!value) return undefined;
+            const day = new Date(value).getDay();
+            const isWeekday = day !== 0 && day !== 6;
+            return isWeekday ? undefined : "Please select a weekday";
+        };
+
+        return (
+            <FinalForm<FormValues>
+                mode="edit"
+                onSubmit={() => {
+                    // not handled
+                }}
+                subscription={{ values: true }}
+            >
+                {() => (
+                    <>
+                        <DatePickerField
+                            name="value"
+                            label="Date Picker"
+                            helperText="Only weekdays are valid"
+                            fullWidth
+                            variant="horizontal"
+                            validate={validateIsWeekday}
+                        />
+
+                        <FinalFormDebug />
+                    </>
+                )}
+            </FinalForm>
+        );
+    },
+};